requirements.txt
asgiref==3.2.7
cffi==1.14.0
cryptography==2.9.2
dj-database-url==0.5.0
Django==3.1.2
django-cors-headers==3.2.1
django-heroku==0.3.1
django-rest-knox==4.1.0
django-restframework==0.0.1
djangorestframework==3.11.0
djangorestframework-expiring-authtoken==0.1.4
gunicorn==20.0.4
mysqlclient==2.0.1
psycopg2==2.8.5
pycparser==2.20
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
whitenoise==5.0.1
